requirements.txt
requests~=2.32.5
python-dotenv~=1.1.1
flet~=0.28.3
plyer~=2.1.0
playwright~=1.56.0